Move profiling-uploader to communications module
