+++ communication/src/test/resources/test-recording.jfr
