+++ dd-java-agent/agent-profiling/profiling-controller/build.gradle
@@ -12,6 +12,8 @@ excludedClassesCoverage += [
   'com.datadog.profiling.controller.ProfilingSystem.StopRecording',
   // This is almost fully abstract class so nothing to test
   'com.datadog.profiling.controller.RecordingData',
+  // A plain enum
+  'com.datadog.profiling.controller.RecordingType',
   // A simple data holder class, nothing to test
   'com.datadog.profiling.controller.ProfilerSettingsSupport',
   'com.datadog.profiling.controller.ProfilingSupport',
