--- dd-java-agent/agent-profiling/profiling-uploader/build.gradle
@@ -1,41 +1,0 @@
-// Set properties before any plugins get loaded
-ext {
-}
-
-apply from: "$rootDir/gradle/java.gradle"
-
-minimumBranchCoverage = 0.80
-
-excludedClassesCoverage += [
-  // Large parts of this class are jvm specific which makes jacoco really confused since we run it only for 'default' jvm
-  'com.datadog.profiling.uploader.util.PidHelper',
-  // The byte counting input/output stream are delegating most of their methods so it does not make sense to force the coverage there
-  'com.datadog.profiling.uploader.ByteCountingInputStream',
-  'com.datadog.profiling.uploader.ByteCountingOutputStream',
-  // A call-back inner class holds few enough instructions to make not testing with different log levels to breach the jacoco limits
-  'com.datadog.profiling.uploader.ProfileUploader.1',
-  // When running on Java 8 without `jfr`, no code is executed
-  'com.datadog.profiling.uploader.util.JfrCliHelper',
-  'com.datadog.profiling.uploader.util.JfrCliHelper.Event'
-]
-
-dependencies {
-  implementation deps.slf4j
-  implementation project(':communication')
-  implementation project(':internal-api')
-  implementation project(':utils:socket-utils')
-  implementation project(':utils:version-utils')
-
-  implementation project(':dd-java-agent:agent-profiling:profiling-controller')
-
-  implementation deps.okhttp
-  implementation group: 'org.lz4', name: 'lz4-java', version: '1.7.1'
-
-  testImplementation deps.junit5
-  testImplementation project(':dd-java-agent:agent-profiling:profiling-testing')
-  testImplementation deps.mockito
-  testImplementation group: 'com.squareup.okhttp3', name: 'mockwebserver', version: versions.okhttp_legacy
-
-  testImplementation(group: 'com.fasterxml.jackson.core', name: 'jackson-databind', version: '2.9.10')
-}
-
