+++ settings.gradle
@@ -55,7 +55,6 @@ include ':dd-java-agent:agent-profiling:profiling-controller-ddprof'
 include ':dd-java-agent:agent-profiling:profiling-controller-openjdk'
 include ':dd-java-agent:agent-profiling:profiling-controller-oracle'
 include ':dd-java-agent:agent-profiling:profiling-testing'
-include ':dd-java-agent:agent-profiling:profiling-uploader'
 include ':dd-java-agent:agent-profiling:profiling-utils'
 
 include ':dd-java-agent:agent-debugger:debugger-bootstrap'
